pyproject.toml
[build-system]
requires = ["setuptools>=61", "setuptools-scm[toml]>=6.2.3"]
build-backend = "setuptools.build_meta"

[project]
name = "milliontrees"
version = "0.3.0"
description = "Benchmark dataset for Airborne Tree Machine Learning"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10,<3.13"
classifiers = [
    "Development Status :: 6 - Mature",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX",
    "Operating System :: Unix",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

authors = [
    { name = "Ben Weinstein", email = "benweinstein@weecology.org" },
    { name = "Henry Senyondo", email = "henrykironde@gmail.com" },
    { name = "Ethan White", email = "ethan@weecology.org" },
]

urls.Documentation = "http://milliontrees.readthedocs.io/en/latest/"
urls.Source = "https://github.com/Weecology/MillionTrees"
urls.Contributors = "https://github.com/Weecology/MillionTrees/graphs/contributors"

dependencies = [
    "numpy",
    "pandas",
    "pillow",
    "torch",
    "torchvision",
    "tqdm",
    "scikit-learn",
    "shapely",
    "albumentations",
    "torchmetrics",
    "pyarrow>=17.0.0",
    "datasets",
    "huggingface_hub",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "bumpversion",
    "yapf",
    "docformatter",
    "black",
    "flake8",
    "label-studio-sdk",
    "paramiko",
    "python-dotenv",
    "outdated",
    "build",
    "twine",
    "deepforest",
    "geopandas",
    "rasterio",
    "scikit-image",
    "requests",
    "neonutilities",
    "dask[diagnostics]",
]
docs = [
    "sphinx",
    "furo",
    "sphinx_markdown_tables",
    "myst_parser",
    "sphinx_rtd_theme",
    "sphinx-design",
    "nbmake",
    "nbsphinx",
    "nbqa",
]
sam3 = [
    "transformers>=4.46.0",
    "accelerate>=0.34.0",
    "safetensors>=0.4.5",
    "matplotlib>=3.8",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages.find = {where = ["src"], include = ["milliontrees*"]}
